Fix test cases.
