+++ test/test-integration/src/test/java/com/alipay/sofa/rpc/server/multi/MultiProtocolServerExpTest.java
@@ -20,7 +20,6 @@ import com.alipay.sofa.rpc.bootstrap.DefaultProviderBootstrap;
 import com.alipay.sofa.rpc.common.RpcConstants;
 import com.alipay.sofa.rpc.config.ProviderConfig;
 import com.alipay.sofa.rpc.config.ServerConfig;
-import com.alipay.sofa.rpc.context.RpcRuntimeContext;
 import com.alipay.sofa.rpc.server.rest.RestService;
 import com.alipay.sofa.rpc.server.rest.RestServiceImpl;
 import com.alipay.sofa.rpc.test.ActivelyDestroyTest;
@@ -29,53 +28,13 @@ import org.junit.Test;
 
 import java.lang.reflect.Field;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * @author <a href="mailto:zhanggeng.zg@antfin.com">GengZhang</a>
  */
-public class MultiProtolServerTest extends ActivelyDestroyTest {
-
-    @Test
-    public void testMultiProtocol() {
-
-        try {
-            // 只有2个线程 执行
-            ServerConfig serverConfig = new ServerConfig()
-                .setStopTimeout(0)
-                .setPort(22222)
-                .setProtocol(RpcConstants.PROTOCOL_TYPE_BOLT)
-                .setQueues(100).setCoreThreads(1).setMaxThreads(2);
-
-            // 发布一个服务，每个请求要执行1秒
-            ProviderConfig<RestService> providerConfig = new ProviderConfig<RestService>()
-                .setInterfaceId(RestService.class.getName())
-                .setRef(new RestServiceImpl())
-                .setServer(serverConfig)
-                .setRepeatedExportLimit(1)
-                .setRegister(false);
-            providerConfig.export();
-
-            ServerConfig serverConfig2 = new ServerConfig()
-                .setStopTimeout(0)
-                .setPort(22223)
-                .setProtocol(RpcConstants.PROTOCOL_TYPE_REST)
-                .setQueues(100).setCoreThreads(1).setMaxThreads(2);
-
-            // 发布一个服务，每个请求要执行1秒
-            ProviderConfig<RestService> providerConfig2 = new ProviderConfig<RestService>()
-                .setInterfaceId(RestService.class.getName())
-                .setRef(new RestServiceImpl())
-                .setServer(serverConfig2)
-                .setRepeatedExportLimit(1)
-                .setRegister(false);
-            providerConfig2.export();
-        } catch (Throwable e) {
-            Assert.fail();
-        }
-
-    }
+public class MultiProtocolServerExpTest extends ActivelyDestroyTest {
 
     @Test
     public void testMultiProtocolExp() throws NoSuchFieldException {
@@ -122,12 +81,12 @@ public class MultiProtolServerTest extends ActivelyDestroyTest {
             Assert.fail();
         } catch (Throwable e) {
             //reflect to fetch export key
-            ConcurrentHashMap<String, AtomicInteger> map = null;
+            ConcurrentMap<String, AtomicInteger> map = null;
 
             Field field = DefaultProviderBootstrap.class.getDeclaredField("EXPORTED_KEYS");
             try {
                 field.setAccessible(true);
-                map = (ConcurrentHashMap<String, AtomicInteger>) field.get(null);
+                map = (ConcurrentMap<String, AtomicInteger>) field.get(null);
             } catch (IllegalAccessException e1) {
                 e1.printStackTrace();
             }
@@ -138,6 +97,5 @@ public class MultiProtolServerTest extends ActivelyDestroyTest {
                 Assert.assertEquals(1, atomicInteger.get());
             }
         }
-
     }
 }
